refactor: Split helper.py into the six subjects it was holding
services/helper/helper.py held 414 lines of six unrelated things: sending to a
Discord channel, writing a user or channel into a message, saying how long ago
something was, working out when a birthday next falls, toggling a role, and
proving a webhook came from Twitch. They shared a filename and nothing else -
the HMAC helpers in particular are a Twitch concern that had been living among
the Discord ones.

Six modules, none over 93 lines: services/send.py, present.py, duration.py,
birthday.py, roles.py, and services/twitch/signature.py where the signature
helpers belong. Bodies were lifted by AST span rather than retyped.
services/__init__.py re-exports the same nineteen names, so every consumer that
went through the facade is untouched; the three that reached into helper.helper
directly now name the module they actually want.

The gate flagged something real on the way past, in a file this only changed an
import line of, and it is fixed here rather than deferred: a Twitch stream title
goes into an embed description wrapped in the bot's own markdown link, and a
title containing "](" ends that link and starts one pointing anywhere. These
alerts are posted for every broadcaster the bot announces, not only the owner.

Worth recording precisely, because I had it wrong: discord.utils.escape_markdown
escapes * _ ~ | and a backtick, and neutralises a WHOLE [label](url) through its
URL awareness - but it leaves a bare bracket alone. That is enough for a field
value, where the attacker would have to supply both brackets, and not enough
inside a link label, where the bot has already supplied the opening one.
_linkable escapes the brackets in one pass for that position; the game name and
the offline title, neither of which sits inside a link, keep escape_markdown.

Verified: the nineteen re-exports all resolve and behave; the thirteen-case
webhook harness is unchanged; four breakout titles no longer escape the link and
an ordinary one still reads as itself.
